1 different results on windows for test dataset (#2)
* More specific output for test

* Change the order of the initializers

* Change int to sizet and remove unused variable

* Cast size_t to int for comparison warning

* Convert weights to size_t

* Change old school cast to static cast

* change everything to size_t for consinstency

* Missed declaration

* Change all points back to double

* Fix some types

* Add some static casts

* Many more static conversions

* Forgotten one definition

* More static casts

* Remove last warnings2

* Add git hasH

* Fix workflow

* Fix workflow

* Syntax of git hash that works for pull and push

---------

+++ bico/_core.cpp
@@ -35,7 +35,7 @@ public:
     virtual ~BicoExternal();
     void addData(double const *array, uint n);
     void addPoint(double const *array);
-    int compute(int *sample_weights,
+    size_t compute(double *sample_weights,
                 double *points);
 
 private:
@@ -66,7 +66,7 @@ void BicoExternal::addPoint(double const *array)
     *_bico << p;
 }
 
-int BicoExternal::compute(int *sample_weights,
+size_t BicoExternal::compute(double *sample_weights,
                           double *points)
 {
     // Retrieve coreset
@@ -82,7 +82,7 @@ int BicoExternal::compute(int *sample_weights,
             points[i * _d + j] = sol->proxysets[0][i][j];
         }
     }
-    int m = sol->proxysets[0].size();
+    size_t m = sol->proxysets[0].size();
     delete sol;
 
     return m;
@@ -123,7 +123,7 @@ extern "C"
 #if defined(_WIN32) || defined(__CYGWIN__)
     __declspec(dllexport)
 #endif
-    int compute(BicoExternal *bico, int *sample_weights,
+    size_t compute(BicoExternal *bico, double *sample_weights,
                 double *points) { return bico->compute(sample_weights, points); }
 
 #if defined(_WIN32) || defined(__CYGWIN__)
